fix typos and code clean
